+++ qcal/__init__.py
@@ -9,4 +9,5 @@ from qcal.gate.gate import Gate
 from qcal.gate.single_qubit import *
 from qcal.gate.two_qubit import *
 from qcal.results import Results
+from qcal.simulation.simulators import StateVectorSimulator
 from qcal.utils import *